+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to the "vscode-maven" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## Unreleased
+- Add context menu on `pom.xml`.
+- Support maven archetype generate.
+
+## Released
 
 ### 0.0.6 
 - Run command in dedicated terminals for each maven project.